Copyedit docs, update microbenchmark
The base checkpoint cost in the microbenchmark has become much slower since the last time I ran it, but that's not really greenback's fault. I don't remember whether the previous run was on the same laptop so it's possible that explains it. I updated the description to be clearer about the conditions.